feat: Deprecate `appendToStream` in favor of `insert` for Event Store Operations (#216)
* .

* Revert "."

This reverts commit b125e251ab0d217e7ad12b964d4a2fc6725cfb60.

* bump

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

* .

+++ core/service/Service/Event/StreamPosition.hs
@@ -6,11 +6,11 @@ module Service.Event.StreamPosition (
 import Core
 
 
-newtype StreamPosition = StreamPosition Int
+newtype StreamPosition = StreamPosition Int64
   deriving (Eq, Show, Ord, Generic)
 
 
-new :: Int -> Maybe StreamPosition
+new :: Int64 -> Maybe StreamPosition
 new position = do
   if position < 0
     then Nothing
